add security configuration

+++ springboot/src/main/java/vn/minhquang/jobhunter/controller/UserController.java
@@ -1,19 +1,20 @@
 package vn.minhquang.jobhunter.controller;
 
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
 import java.util.List;
 
 import vn.minhquang.jobhunter.domain.User;
 import vn.minhquang.jobhunter.service.UserService;
+import vn.minhquang.jobhunter.service.error.IdInvalidException;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
@@ -21,13 +22,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 public class UserController {
   
   private final UserService userService;
+  private final PasswordEncoder passwordEncoder;
 
-  UserController(UserService userService) {
+  UserController(UserService userService, PasswordEncoder passwordEncoder) {
+    this.passwordEncoder = passwordEncoder;
     this.userService = userService;
   }
 
   @PostMapping("/users")
   public ResponseEntity<User> createUser(@RequestBody User user) {
+    String encodedPassword = this.passwordEncoder.encode(user.getPassword());
+    user.setPassword(encodedPassword);
+
     User createdUser = this.userService.handleCreateUser(user);
     return ResponseEntity.status(HttpStatus.CREATED).body(createdUser);
   }
@@ -51,7 +57,13 @@ public class UserController {
   }
   
   @DeleteMapping("/users/{id}")
-  public ResponseEntity<String> deleteUser(@PathVariable("id") Long id) {
+  public ResponseEntity<String> deleteUser(@PathVariable("id") Long id) 
+  throws IdInvalidException {
+
+    if (id <= 0) {
+      throw new IdInvalidException("Id must be greater than 0");
+    }
+
     this.userService.deleteUser(id);
     return ResponseEntity.ok().body("User deleted successfully");
   }
